+++ lib/llm/src/block_manager/block/transfer.rs
@@ -397,7 +397,6 @@ where
         TransferStrategy::CudaAsyncH2D
         | TransferStrategy::CudaAsyncD2H
         | TransferStrategy::CudaAsyncD2D => {
-
             if RB::write_to_strategy() == TransferStrategy::CudaAsyncH2D {
                 // Verify H2D: Host -> Device transfer
                 if sources.len() > 0 {
@@ -441,320 +440,7 @@ where
                     }
                 }
 
-                // Use custom copy_kernel_KV for multiple H2D blocks only
-                if sources.len() > 1 {
-                    // Check if blocks are fully contiguous (required for our simple kernel)
-                    let src_data = sources[0].block_data();
-                    let is_fully_contiguous = src_data.is_fully_contiguous();
-
-                    if is_fully_contiguous {
-                        tracing::info!("Using copy_kernel_KV (FATBIN) for {} H2D blocks [FullyContiguous]", sources.len());
-
-                        let block_size = src_data.block_view()?.size();
-                        let mut src_ptrs = Vec::with_capacity(sources.len());
-                        let mut dst_ptrs = Vec::with_capacity(sources.len());
-
-                        for (src, dst) in sources.iter().zip(targets.iter()) {
-                            let src_view = src.block_data().block_view()?;
-                            let dst_view = dst.block_data().block_view()?;
-
-                            unsafe {
-                                src_ptrs.push(src_view.as_ptr() as u64);
-                                dst_ptrs.push(dst_view.as_ptr() as u64);
-                            }
-                        }
-
-                        tracing::info!("Copy size: {} bytes, {} pairs [FullyContiguous blocks]", block_size, sources.len());
-
-                        // For now, skip pooled buffers to avoid CUDA context issues during initialization
-                        // TODO: Implement proper CUDA context-aware buffer pool initialization
-                        let use_pooled = false; // sources.len() <= MAX_PREALLOCATED_BLOCKS;
-                        let buffer_guard = None::<TransferBufferGuard>;
-
-                    if let Some(ref guard) = buffer_guard {
-                        // Use pooled buffer - fast path
-                        let buffer = guard.get();
-
-                        unsafe {
-                            // Bind CUDA context to current thread for pooled operations
-                            let _context_guard = ctx.stream().context().bind_to_thread();
-
-                            // Copy pointer arrays to pooled device memory
-                            cuda_result::memcpy_htod_async(buffer.src_ptrs, &src_ptrs, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy source pointers: {}", e)))?;
-                            cuda_result::memcpy_htod_async(buffer.dst_ptrs, &dst_ptrs, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy dest pointers: {}", e)))?;
-
-                            // Launch kernel with pooled buffers
-                            let kernel = get_copy_kernel()?;
-
-                            // Optimal grid sizing for grid-stride kernel
-                            let threads_per_block = 256u32;
-                            let max_blocks = 1024u32; // Reasonable limit, let grid-stride handle the rest
-                            let blocks_needed = std::cmp::min(max_blocks, sources.len() as u32);
-
-                            let grid_dim = (blocks_needed, 1, 1);
-                            let block_dim = (threads_per_block, 1, 1);
-
-                            // cuLaunchKernel expects pointers to parameter values
-                            let src_ptr_param = buffer.src_ptrs;
-                            let dst_ptr_param = buffer.dst_ptrs;
-                            let size_param = block_size;
-                            let num_pairs_param = sources.len() as i32; // Match kernel's int num_pairs
-
-                            let params = [
-                                &src_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &dst_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &size_param as *const _ as *mut std::ffi::c_void,       // Pointer to size value
-                                &num_pairs_param as *const _ as *mut std::ffi::c_void,  // Pointer to num_pairs (int)
-                            ];
-
-                            let result = cudarc::driver::sys::cuLaunchKernel(
-                                kernel,
-                                grid_dim.0, grid_dim.1, grid_dim.2,  // grid dimensions
-                                block_dim.0, block_dim.1, block_dim.2,  // block dimensions
-                                0, // shared memory
-                                ctx.stream().cu_stream(),
-                                params.as_ptr() as *mut *mut std::ffi::c_void,
-                                std::ptr::null_mut(), // extra
-                            );
-                            if result != cudarc::driver::sys::cudaError_enum::CUDA_SUCCESS {
-                                return Err(TransferError::ExecutionError(format!("Kernel launch failed: {:?}", result)));
-                            }
-                        }
-
-                        let pool = get_transfer_buffer_pool()?;
-                        tracing::info!("copy_kernel_KV launched with pooled buffer {} (available: {}/{})",
-                                 buffer.id, pool.len(), pool.total_buffers);
-                    } else {
-                        // Fall back to malloc/memcpy/free approach using pinned host memory
-                        // Note: This path has potential memory leak due to async completion
-                        if !use_pooled {
-                            tracing::info!("Block count {} exceeds buffer capacity {}, using malloc fallback",
-                                     sources.len(), MAX_PREALLOCATED_BLOCKS);
-                        } else {
-                            tracing::info!("No pooled buffers available, using malloc fallback");
-                        }
-
-                        // Collect all source and destination addresses for each layer's K/V components
-                        let mut src_addresses = Vec::new();
-                        let mut dst_addresses = Vec::new();
-
-                        let num_layers = sources[0].block_data().num_layers();
-                        for (src_block, dst_block) in sources.iter().zip(targets.iter()) {
-                            for layer_idx in 0..num_layers {
-                                // K cache addresses (outer_dim = 0)
-                                if let (Ok(src_k), Ok(dst_k)) = (
-                                    src_block.block_data().layer_view(layer_idx, 0),
-                                    dst_block.block_data().layer_view(layer_idx, 0)
-                                ) {
-                                    unsafe {
-                                        src_addresses.push(src_k.as_ptr() as u64);
-                                        dst_addresses.push(dst_k.as_ptr() as u64);
-                                    }
-                                }
-                                // V cache addresses (outer_dim = 1)
-                                if let (Ok(src_v), Ok(dst_v)) = (
-                                    src_block.block_data().layer_view(layer_idx, 1),
-                                    dst_block.block_data().layer_view(layer_idx, 1)
-                                ) {
-                                    unsafe {
-                                        src_addresses.push(src_v.as_ptr() as u64);
-                                        dst_addresses.push(dst_v.as_ptr() as u64);
-                                    }
-                                }
-                            }
-                        }
-
-                        let layer_size = sources[0].block_data().layer_view(0, 0)?.size();
-                        tracing::info!("Copy size: {} bytes, {} transfers [FullyContiguous fallback: {} src/dst pairs]",
-                                 layer_size, src_addresses.len(), src_addresses.len());
-
-                        let ptr_array_size = src_addresses.len() * std::mem::size_of::<u64>();
-                        let mut d_src_ptrs = 0u64;
-                        let mut d_dst_ptrs = 0u64;
-
-                        unsafe {
-                            // Bind CUDA context to current thread for allocations
-                            let _context_guard = ctx.stream().context().bind_to_thread();
-
-                            // Allocate device memory for pointer arrays
-                            d_src_ptrs = cuda_result::malloc_sync(ptr_array_size)
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to allocate device source pointers: {}", e)))?;
-                            d_dst_ptrs = cuda_result::malloc_sync(ptr_array_size)
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to allocate device dest pointers: {}", e)))?;
-
-                            // Copy pointer arrays to device (direct from Vec)
-                            cuda_result::memcpy_htod_async(d_src_ptrs, &src_addresses, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy source pointers: {}", e)))?;
-                            cuda_result::memcpy_htod_async(d_dst_ptrs, &dst_addresses, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy dest pointers: {}", e)))?;
-
-                            // Launch kernel using fallback malloc approach
-                            let kernel = get_copy_kernel()?;
-
-                            // Optimal grid sizing for grid-stride kernel (not one block per transfer!)
-                            let threads_per_block = 256u32;
-                            let max_blocks = 1024u32; // Reasonable limit, let grid-stride handle the rest
-                            let blocks_needed = std::cmp::min(max_blocks, src_addresses.len() as u32);
-
-                            let grid_dim = (blocks_needed, 1, 1);
-                            let block_dim = (threads_per_block, 1, 1);
-
-                            tracing::info!("Grid-stride kernel: {} pairs, {} blocks × {} threads (grid-stride handles overflow)",
-                                     src_addresses.len(), blocks_needed, threads_per_block);
-
-                            // cuLaunchKernel expects pointers to parameter values
-                            let src_ptr_param = d_src_ptrs;
-                            let dst_ptr_param = d_dst_ptrs;
-                            let size_param = layer_size;  // Size per layer component (32KB)
-                            let num_pairs_param = src_addresses.len() as i32; // Match kernel's int num_pairs
-
-                            let params = [
-                                &src_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &dst_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &size_param as *const _ as *mut std::ffi::c_void,       // Pointer to size value
-                                &num_pairs_param as *const _ as *mut std::ffi::c_void,  // Pointer to num_pairs (int)
-                            ];
-
-                            let result = cudarc::driver::sys::cuLaunchKernel(
-                                kernel,
-                                grid_dim.0, grid_dim.1, grid_dim.2,  // grid dimensions
-                                block_dim.0, block_dim.1, block_dim.2,  // block dimensions
-                                0, // shared memory
-                                ctx.stream().cu_stream(),
-                                params.as_ptr() as *mut *mut std::ffi::c_void,
-                                std::ptr::null_mut(), // extra
-                            );
-                            if result != cudarc::driver::sys::cudaError_enum::CUDA_SUCCESS {
-                                return Err(TransferError::ExecutionError(format!("FullyContiguous kernel launch failed: {:?}", result)));
-                            }
-
-                            // Note: Memory cleanup deferred - will be cleaned up when stream completes
-                            // TODO: Implement proper async cleanup or use pooled buffers to avoid this issue
-                            // let _ = cuda_result::free_sync(d_src_ptrs);
-                            // let _ = cuda_result::free_sync(d_dst_ptrs);
-                        }
-
-                        tracing::info!("copy_kernel_KV launched successfully (FullyContiguous malloc fallback)");
-                    }
-                    } else {
-                        // LayerSeparate layout detected - collect all layer addresses for scatter kernel
-                        let num_layers = src_data.num_layers();
-                        let num_outer_dims = src_data.num_outer_dims();
-                        let layer_size = src_data.layer_view(0, 0)?.size();
-                        let total_size = layer_size * num_layers * num_outer_dims;
-
-                        tracing::info!("Using copy_kernel_KV (FATBIN) for {} LayerSeparate blocks [{}L×{}O×{}B]",
-                                 sources.len(), num_layers, num_outer_dims, layer_size);
-
-                        // Collect all source and destination addresses for each layer's K/V components
-                        let mut src_addresses = Vec::new();
-                        let mut dst_addresses = Vec::new();
-
-                        for (src_block, dst_block) in sources.iter().zip(targets.iter()) {
-                            for layer_idx in 0..num_layers {
-                                // K cache addresses (outer_dim = 0)
-                                if let (Ok(src_k), Ok(dst_k)) = (
-                                    src_block.block_data().layer_view(layer_idx, 0),
-                                    dst_block.block_data().layer_view(layer_idx, 0)
-                                ) {
-                                    unsafe {
-                                        src_addresses.push(src_k.as_ptr() as u64);
-                                        dst_addresses.push(dst_k.as_ptr() as u64);
-                                    }
-                                }
-                                // V cache addresses (outer_dim = 1)
-                                if let (Ok(src_v), Ok(dst_v)) = (
-                                    src_block.block_data().layer_view(layer_idx, 1),
-                                    dst_block.block_data().layer_view(layer_idx, 1)
-                                ) {
-                                    unsafe {
-                                        src_addresses.push(src_v.as_ptr() as u64);
-                                        dst_addresses.push(dst_v.as_ptr() as u64);
-                                    }
-                                }
-                            }
-                        }
-
-                        tracing::info!("Copy size: {} bytes, {} transfers [LayerSeparate: {} src/dst pairs]",
-                                 layer_size, src_addresses.len(), src_addresses.len());
-
-                        // Launch kernel with collected LayerSeparate addresses
-                        let ptr_array_size = src_addresses.len() * std::mem::size_of::<u64>();
-                        let mut d_src_ptrs = 0u64;
-                        let mut d_dst_ptrs = 0u64;
-
-                        unsafe {
-                            // Bind CUDA context to current thread for allocations
-                            let _context_guard = ctx.stream().context().bind_to_thread();
-
-                            // Allocate device memory for pointer arrays
-                            d_src_ptrs = cuda_result::malloc_sync(ptr_array_size)
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to allocate LayerSeparate source pointers: {}", e)))?;
-                            d_dst_ptrs = cuda_result::malloc_sync(ptr_array_size)
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to allocate LayerSeparate dest pointers: {}", e)))?;
-
-                            // Copy pointer arrays to device
-                            cuda_result::memcpy_htod_async(d_src_ptrs, &src_addresses, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy LayerSeparate source pointers: {}", e)))?;
-                            cuda_result::memcpy_htod_async(d_dst_ptrs, &dst_addresses, ctx.stream().cu_stream())
-                                .map_err(|e| TransferError::ExecutionError(format!("Failed to copy LayerSeparate dest pointers: {}", e)))?;
-
-                            // Launch kernel for LayerSeparate transfers
-                            let kernel = get_copy_kernel()?;
-
-                            // Optimal grid sizing for grid-stride kernel (not one block per transfer!)
-                            let threads_per_block = 256u32;
-                            let max_blocks = 1024u32; // Reasonable limit, let grid-stride handle the rest
-                            let blocks_needed = std::cmp::min(max_blocks, src_addresses.len() as u32);
-
-                            let grid_dim = (blocks_needed, 1, 1);
-                            let block_dim = (threads_per_block, 1, 1);
-
-                            tracing::info!(" LayerSeparate grid-stride: {} pairs, {} blocks × {} threads",
-                                     src_addresses.len(), blocks_needed, threads_per_block);
-
-                            // cuLaunchKernel expects pointers to parameter values
-                            let src_ptr_param = d_src_ptrs;
-                            let dst_ptr_param = d_dst_ptrs;
-                            let size_param = layer_size;  // Size per layer component (32KB)
-                            let num_pairs_param = src_addresses.len() as i32; // Match kernel's int num_pairs
-
-                            let params = [
-                                &src_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &dst_ptr_param as *const _ as *mut std::ffi::c_void,    // Pointer to device address
-                                &size_param as *const _ as *mut std::ffi::c_void,       // Pointer to size value
-                                &num_pairs_param as *const _ as *mut std::ffi::c_void,  // Pointer to num_pairs (int)
-                            ];
-
-                            let result = cudarc::driver::sys::cuLaunchKernel(
-                                kernel,
-                                grid_dim.0, grid_dim.1, grid_dim.2,  // grid dimensions
-                                block_dim.0, block_dim.1, block_dim.2,  // block dimensions
-                                0, // shared memory
-                                ctx.stream().cu_stream(),
-                                params.as_ptr() as *mut *mut std::ffi::c_void,
-                                std::ptr::null_mut(), // extra
-                            );
-                            if result != cudarc::driver::sys::cudaError_enum::CUDA_SUCCESS {
-                                return Err(TransferError::ExecutionError(format!("LayerSeparate kernel launch failed: {:?}", result)));
-                            }
-
-                            // Note: Memory cleanup deferred - will be cleaned up when stream completes
-                            // TODO: Implement proper async cleanup or use pooled buffers to avoid this issue
-                            // let _ = cuda_result::free_sync(d_src_ptrs);
-                            // let _ = cuda_result::free_sync(d_dst_ptrs);
-                        }
-
-                        tracing::info!("copy_kernel_KV launched successfully for LayerSeparate layout");
-                    }
-                } else {
-                    // Fall back to individual copy for single H2D blocks
-                    for (src, dst) in sources.iter().zip(targets.iter_mut()) {
-                        cuda::copy_block(src, dst, ctx.stream().as_ref(), RB::write_to_strategy())?;
-                    }
-                }
+                cuda::copy_blocks_with_customized_kernel(sources, targets, ctx.stream().as_ref(), RB::write_to_strategy())?;
             } else if RB::write_to_strategy() == TransferStrategy::CudaAsyncD2H {
                 // Verify D2H: Device -> Host transfer
                 if sources.len() > 0 {
@@ -798,19 +484,12 @@ where
                     }
                 }
 
-                // D2H transfers use standard block copy
-                tracing::info!("ziqif copy_blocks");
-                cuda::copy_blocks(sources, targets, ctx.stream().as_ref(), RB::write_to_strategy())?;
-                // for (src, dst) in sources.iter().zip(targets.iter_mut()) {
-                //     cuda::copy_block(src, dst, ctx.stream().as_ref(), RB::write_to_strategy())?;
-                // }
+                cuda::copy_blocks_with_customized_kernel(sources, targets, ctx.stream().as_ref(), RB::write_to_strategy())?;
             } else {
-                // For all other cases (D2D, etc.), use standard block copy
-                tracing::info!("ziqif copy_blocks");
-                cuda::copy_blocks(sources, targets, ctx.stream().as_ref(), RB::write_to_strategy())?;
-                // for (src, dst) in sources.iter().zip(targets.iter_mut()) {
-                //     cuda::copy_block(src, dst, ctx.stream().as_ref(), RB::write_to_strategy())?;
-                // }
+                // D2D transfers use standard block copy
+                for (src, dst) in sources.iter().zip(targets.iter_mut()) {
+                    cuda::copy_block(src, dst, ctx.stream().as_ref(), RB::write_to_strategy())?;
+                }
             }
 
             // todo: acquire an cuda event, recored on the stream, drop the stream, await on the event.
